requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.49
python-multipart==0.0.9
reportlab==4.1.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
aiofiles==23.2.1
python-dotenv==1.0.0
aiosmtplib==3.0.1
email-validator==2.1.0
